+++ iTunesSearch/SearchResultsTableViewController.swift
@@ -1,0 +1,44 @@
+//
+//  SearchResultsTableViewController.swift
+//  iTunesSearch
+//
+//  Created by Bhawnish Kumar on 3/10/20.
+//  Copyright © 2020 Bhawnish Kumar. All rights reserved.
+//
+
+import UIKit
+
+class SearchResultsTableViewController: UITableViewController {
+   
+    @IBOutlet weak var searchBar: UISearchBar!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+
+    @IBAction func segmentedControl(_ sender: Any) {
+    }
+    
+var searchResultController = SearchResultController()
+    
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        // #warning Incomplete implementation, return the number of rows
+        return searchResultController.searchResults.count
+    }
+
+    
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "iTunesCell", for: indexPath)
+        let result = searchResultController.searchResults[indexPath.row]
+        cell.textLabel?.text = result.title
+        cell.detailTextLabel?.text = result.creator
+        // Configure the cell...
+
+        return cell
+    }
+    
+
+   
+}
+
